requirements.txt
gitpython>=3.1.0
requests>=2.25.1
python-dotenv>=0.19.0
pyyaml>=5.4.1
beautifulsoup4>=4.9.3
click>=8.0.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.5b2
isort>=5.9.1
mypy>=0.910
pylint>=2.8.2
PyQt6>=6.2.2
